ingestion/requirements.txt
feedparser
newspaper3k
python-dotenv
loguru
lxml[html_clean]
pymongo
